dom/base/test/jsmodules/importmaps/test_dynamic_import_reject_importMap.html

<!DOCTYPE html>
<head>
  <meta charset=utf-8>
  <title>Test import map should be rejected.</title>
</head>
<body onload='testLoaded()'>
<script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

<!--Note: This test should be replaced by-->
<!--test_importMap_between_dynamic_imports.html when the pref-->
<!--"dom.multiple_import_maps.enabled" is turned on.-->

<!--When the multiple_import_maps.enabled is false:-->
<!--There is a dynamic import before the import map tag, so the import map-->
<!--cannot be accepted according to the spec.-->
<!--And because the import map is rejected, so the module specifier-->
<!--"./scope1/module_simpleExport.mjs" won't be remapped to-->
<!--"./module_simpleExport.mjs".-->

<script>
  import("./module_simpleExport.mjs");
</script>

<script type="importmap" onerror='importMapError()'>
{
  "imports": {
    "./scope1/module_simpleExport.mjs": "./module_simpleExport.mjs"
  }
}
</script>

<script>
  SimpleTest.waitForExplicitFinish();

  let hasError = false;
  // eslint-disable-next-line no-unused-vars
  function importMapError() {
    hasError = true;
  }

  // If the multiple import maps is supported, the import map will be accepted,
  // otherwise it is ignored.
  const pref = SpecialPowers.getBoolPref("dom.multiple_import_maps.enabled");
  const expectedValue = pref ? 42 : 84;
  const onerrorExpected = !pref;

  // eslint-disable-next-line no-unused-vars
  function testLoaded() {
    import("./scope1/module_simpleExport.mjs").then((ns) => {
      ok(ns.x == expectedValue, 'Check simple imported value result: ' + ns.x);
      ok(hasError == onerrorExpected, "onerror of the import map should be called.");
    }).catch((e) => {
      ok(false, "throws " + e);
    }).then(() => {
      SimpleTest.finish();
    });
  }
</script>
</body>
